+++ apps/api/env.d.ts
@@ -14,6 +14,7 @@ declare module "process" {
           GOOGLE_CLIENT_ID: string;
           GOOGLE_CLIENT_SECRET: string;
           GOOGLE_REFRESH_TOKEN: string;
+          EMAIL_USER: string;
         };
       }
     }
