+++ turbo.json
@@ -13,7 +13,8 @@
         "JWT_SECRET",
         "GOOGLE_CLIENT_ID",
         "GOOGLE_CLIENT_SECRET",
-        "GOOGLE_REFRESH_TOKEN"
+        "GOOGLE_REFRESH_TOKEN",
+        "EMAIL_USER"
       ]
     },
     "@spin-spot/client": {
